requirements.txt
pandas
matplotlib
seaborn
emoji
wordcloud
vaderSentiment
urlextract
instaloader
flask
google-generativeai
python-dotenv
gunicorn
